Use UTF-8 for import and export
Right now, the plugin is using the platform default encoding, which is UTF-8 for Linux and Mac and Windows-1252 for Windows. In today's world, using UTF-8 for all three is the sensible choice.

This enables using characters such as emojis reliably. It also enables sharing the exported file with someone who uses a different OS.

The Runelite version was also bumped from .15 to .21.